# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Laravel Deep Challenge API
  version: 1.0.0
  description: API documentation for the Laravel Deep Challenge backend.
servers:
  - url: http://localhost:8000
    description: Local Development Server
paths:
  /profile:
    get:
      summary: Get User Profile
      description: Returns the authenticated user's profile information.
      security:
        - sessionAuth: []
        - sanctum: []
      responses:
        '200':
          description: Successful operation
          content:
            text/html:
              schema:
                type: string
                description: Returns the profile edit view (since this is a web route).
    patch:
      summary: Update User Profile
      description: Updates the authenticated user's profile.
      security:
        - sessionAuth: []
        - sanctum: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: User's name
                  example: John Doe
                email:
                  type: string
                  format: email
                  description: User's email
                  example: john@example.com
                password:
                  type: string
                  format: password
                  description: New password (optional)
                  example: newpassword123
                password_confirmation:
                  type: string
                  format: password
                  description: Confirm new password
                  example: newpassword123
                profile_image:
                  type: string
                  format: binary
                  description: Profile image file (jpg, jpeg, png)
              required:
                - name
                - email
      responses:
        '302':
          description: Redirects back to profile page on success.
        '422':
          description: Validation error.
    delete:
      summary: Delete User Account
      description: Deletes the authenticated user's account.
      security:
        - sessionAuth: []
        - sanctum: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                  description: Current password for confirmation
                  example: password
              required:
                - password
      responses:
        '302':
          description: Redirects to homepage on success.
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: XSRF-TOKEN
      description: standard Laravel session authentication.
    sanctum: 
      type: http
      scheme: bearer